==> 01_topic_modeling/01_data_loading.ipynb <==
import os

# please download the data using the link in README
if not os.path.exists('data/aylien_covid_news_data.tar.gz'):
    print('Please download the data first.')
    exit()

if not os.path.exists('data/aylien_covid_news_data.jsonl'):
    os.system('tar -xvf data/aylien_covid_news_data.tar.gz')
    os.system('rm README.txt')

import jsonlines
import pandas as pd
import numpy as np

# select the news articles from only these six sources

target_sources = ['CNN', 'Fox News', 'Huffington Post US', 
                  'Breitbart', 'The New York Times', 'New York Post']

aylien_json_path = 'aylien_covid_news_data.jsonl'
articles = []
i = 0
with jsonlines.open(aylien_json_path) as reader:
    for obj in reader:
        source =  obj['source']['name']
        if source in target_sources:
            summary = obj['summary']['sentences']
            country = 'US'
            info = (obj['title'], source, obj['published_at'], obj['language'], country, tuple(summary))
            articles.append(info)
            i += 1
            if i % 2000 == 0:
                print(i, source)
data_raw = pd.DataFrame(articles, columns=['title', 'source', 'date', 'language', 'country', 'summary'])

import copy
data = copy.deepcopy(data_raw)
data.head()

data.shape

# rename the sources
source_map = {'CNN': 'cnn',
             'Fox News': 'fox',
             'Huffington Post US':'huff',
             'Breitbart': 'breit',
             'The New York Times': 'nyt',
             'New York Post': 'nyp'}
data['source'] = data['source'].map(source_map)

data.head()

# drop the duplicate news
data = data.drop_duplicates(subset=['title']).drop_duplicates(subset=['summary'])

# normalize the text
def prepocess_summary(a):
    import unicodedata, re
    a = ' '.join(a)
    a = re.sub(' +', ' ', a)
    a = a.replace('\n', ' ')
    a = unicodedata.normalize("NFKD", a)
    return a
data['summary'] = data['summary'].apply(prepocess_summary)

# drop the rows with null entries
data = data[data['summary'] != '']
data = data[data['title'] != '']

data.head()

# index the articles
data.insert(0, 'idx', np.arange(data.shape[0]))
data.head()

data['date'] = pd.to_datetime(data['date'])
data['month'] = data['date'].dt.month
data['week'] = data['date'].dt.week
data.head()

# save the news dataframe
os.makedirs('data', exist_ok=True)
data.to_csv('data/df_news.csv', index=False)

# count the # of articles in each month
from collections import Counter
Counter(data['month'])

# count the # of articles from each source
Counter(data['source'])

# count the # of articls of different sources in different mon
for source in ['cnn','fox','huff','breit','nyt','nyp']:
    for month in [1,2,3,4,5,6,7]:
        print(source, month, ((data['source']==source) & (data['month']==month)).sum())
